+++ fbpcs/common/tests/repository/test_instance_local.py
@@ -9,48 +9,32 @@ import unittest
 from pathlib import Path
 from unittest.mock import MagicMock, mock_open, patch
 
-from fbpcs.common.entity.pcs_mpc_instance import PCSMPCInstance
+from fbpcs.common.entity.stage_state_instance import StageStateInstance
 from fbpcs.common.repository.instance_local import LocalInstanceRepository
-
-from fbpcs.private_computation.service.mpc.entity.mpc_instance import (
-    MPCInstanceStatus,
-    MPCParty,
-)
 
 TEST_BASE_DIR = Path("./")
 TEST_INSTANCE_ID = "test-instance-id"
-TEST_GAME_NAME = "lift"
-TEST_MPC_PARTY = MPCParty.SERVER
-TEST_NUM_WORKERS = 1
-TEST_SERVER_IPS = ["192.0.2.0"]
-TEST_GAME_ARGS = [{}]
 ERROR_MSG_ALREADY_EXISTS = f"{TEST_INSTANCE_ID} already exists"
 ERROR_MSG_NOT_EXISTS = f"{TEST_INSTANCE_ID} does not exist"
 
 
 class TestLocalInstanceRepository(unittest.TestCase):
     def setUp(self):
-        self.mpc_instance = PCSMPCInstance.create_instance(
-            instance_id=TEST_INSTANCE_ID,
-            game_name=TEST_GAME_NAME,
-            mpc_party=TEST_MPC_PARTY,
-            num_workers=TEST_NUM_WORKERS,
-            server_ips=TEST_SERVER_IPS,
-            status=MPCInstanceStatus.CREATED,
-            game_args=TEST_GAME_ARGS,
+        self.instance = StageStateInstance(
+            instance_id=TEST_INSTANCE_ID, stage_name="compute"
         )
         self.local_instance_repo = LocalInstanceRepository(TEST_BASE_DIR)
 
     def test_create_existing_instance(self):
         self.local_instance_repo._exist = MagicMock(return_value=True)
         with self.assertRaisesRegex(RuntimeError, ERROR_MSG_ALREADY_EXISTS):
-            self.local_instance_repo.create(self.mpc_instance)
+            self.local_instance_repo.create(self.instance)
 
     @patch("builtins.open")
     def test_create_non_existing_instance(self, mock_open):
         self.local_instance_repo._exist = MagicMock(return_value=False)
         path = TEST_BASE_DIR.joinpath(TEST_INSTANCE_ID)
-        self.assertIsNone(self.local_instance_repo.create(self.mpc_instance))
+        self.assertIsNone(self.local_instance_repo.create(self.instance))
         mock_open.assert_called_with(path, "w")
 
     def test_read_non_existing_instance(self):
@@ -60,28 +44,28 @@ class TestLocalInstanceRepository(unittest.TestCase):
 
     def test_read_existing_instance(self):
         self.local_instance_repo._exist = MagicMock(return_value=True)
-        data = self.mpc_instance.dumps_schema()
+        data = self.instance.dumps_schema()
         path = TEST_BASE_DIR.joinpath(TEST_INSTANCE_ID)
         with patch("builtins.open", mock_open(read_data=data)) as mock_file:
             self.assertEqual(open(path).read().strip(), data)
-            mpc_instance = PCSMPCInstance.loads_schema(
+            instance = StageStateInstance.loads_schema(
                 self.local_instance_repo.read(TEST_INSTANCE_ID)
             )
-            self.assertEqual(self.mpc_instance, mpc_instance)
+            self.assertEqual(self.instance, instance)
             mock_file.assert_called_with(path, "r")
 
     def test_update_non_existing_instance(self):
         self.local_instance_repo._exist = MagicMock(return_value=False)
         with self.assertRaisesRegex(RuntimeError, ERROR_MSG_NOT_EXISTS):
-            self.local_instance_repo.update(self.mpc_instance)
+            self.local_instance_repo.update(self.instance)
 
     @patch("builtins.open")
     def test_update_existing_instance(self, mock_open):
         self.local_instance_repo._exist = MagicMock(return_value=True)
-        new_mpc_instance = copy.deepcopy(self.mpc_instance)
-        new_mpc_instance.game_name = "aggregator"
+        new_instance = copy.deepcopy(self.instance)
+        new_instance.stage_name = "aggregate"
         path = TEST_BASE_DIR.joinpath(TEST_INSTANCE_ID)
-        self.assertIsNone(self.local_instance_repo.update(new_mpc_instance))
+        self.assertIsNone(self.local_instance_repo.update(new_instance))
         mock_open.assert_called_with(path, "w")
 
     def test_exists(self):